backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-multipart
python-dotenv
pydantic
email-validator